reshuffle slow tests to integration, fix kraken integration tests
