+++ test/integration/test_assembly.py
@@ -47,3 +47,45 @@ class TestAssemble(TestCaseWithTmp):
 
 # in order to test the actual de novo pipeline, we need to add a clip db for trimmomatic
 # then we should test from G5012.3.testreads.bam all the way through the assembly pipe
+
+class TestRefineAssembly(TestCaseWithTmp):
+    def test_ebov_refine1(self):
+        inDir = util.file.get_test_input_path(self)
+        inFasta = os.path.join(inDir, 'impute.ebov.fasta')
+        imputeFasta = util.file.mkstempfname('.imputed.fasta')
+        refine1Fasta = util.file.mkstempfname('.refine1.fasta')
+        shutil.copy(inFasta, imputeFasta)
+        tools.picard.CreateSequenceDictionaryTool().execute(imputeFasta, overwrite=True)
+        tools.novoalign.NovoalignTool().index_fasta(imputeFasta)
+        assembly.refine_assembly(
+            imputeFasta,
+            os.path.join(util.file.get_test_input_path(), 'G5012.3.testreads.bam'),
+            refine1Fasta,
+            # normally -r Random, but for unit tests, we want deterministic behavior
+            novo_params='-r None -l 30 -x 20 -t 502',
+            min_coverage=2,
+            threads=4)
+        actual = str(Bio.SeqIO.read(refine1Fasta, 'fasta').seq)
+        expected = str(Bio.SeqIO.read(os.path.join(inDir, 'expected.ebov.refine1.fasta'), 'fasta').seq)
+        self.assertEqual(actual, expected)
+
+    def test_ebov_refine2(self):
+        inDir = util.file.get_test_input_path(self)
+        inFasta = os.path.join(inDir, 'expected.ebov.refine1.fasta')
+        refine1Fasta = util.file.mkstempfname('.refine1.fasta')
+        refine2Fasta = util.file.mkstempfname('.refine2.fasta')
+        shutil.copy(inFasta, refine1Fasta)
+        tools.picard.CreateSequenceDictionaryTool().execute(refine1Fasta, overwrite=True)
+        tools.novoalign.NovoalignTool().index_fasta(refine1Fasta)
+        assembly.refine_assembly(
+            refine1Fasta,
+            os.path.join(util.file.get_test_input_path(), 'G5012.3.testreads.bam'),
+            refine2Fasta,
+            # normally -r Random, but for unit tests, we want deterministic behavior
+            novo_params='-r None -l 40 -x 20 -t 100',
+            min_coverage=3,
+            threads=4)
+        actual = str(Bio.SeqIO.read(refine2Fasta, 'fasta').seq)
+        expected = str(Bio.SeqIO.read(os.path.join(inDir, 'expected.ebov.refine2.fasta'), 'fasta').seq)
+        self.assertEqual(actual, expected)
+
